use setup class so we do not rely on reader prototype
