+++ extension/src/test/suite/extension.test.ts
@@ -2,17 +2,15 @@ import { join, resolve } from 'path'
 import { afterEach, beforeEach, describe, it, suite } from 'mocha'
 import { expect } from 'chai'
 import { stub, restore, spy, match } from 'sinon'
-import { window, commands, workspace, Uri } from 'vscode'
+import { window, commands, workspace } from 'vscode'
 import {
   closeAllEditors,
-  configurationChangeEvent,
   mockDisposable,
   mockDuration,
   quickPickInitialized,
   selectQuickPickItem
 } from './util'
 import { mockHasCheckpoints } from './experiments/util'
-import { WEBVIEW_TEST_TIMEOUT } from './timeouts'
 import { Disposable } from '../../extension'
 import * as Python from '../../extensions/python'
 import { DvcReader } from '../../cli/dvc/reader'
@@ -20,19 +18,12 @@ import expShowFixture from '../fixtures/expShow/base/output'
 import plotsDiffFixture from '../fixtures/plotsDiff/output'
 import * as Disposer from '../../util/disposable'
 import { RegisteredCommands } from '../../commands/external'
-import * as Setup from '../../setup/runner'
-import * as Watcher from '../../fileSystem/watcher'
 import * as Telemetry from '../../telemetry'
 import { EventName } from '../../telemetry/constants'
 import { OutputChannel } from '../../vscode/outputChannel'
 import { WorkspaceExperiments } from '../../experiments/workspace'
-import { QuickPickItemWithValue } from '../../vscode/quickPick'
 import { GitReader } from '../../cli/git/reader'
-import { Config } from '../../config'
-import {
-  EXPERIMENT_WORKSPACE_ID,
-  MIN_CLI_VERSION
-} from '../../cli/dvc/contract'
+import { MIN_CLI_VERSION } from '../../cli/dvc/contract'
 import { ConfigKey, setConfigValue } from '../../vscode/config'
 
 suite('Extension Test Suite', () => {
@@ -55,55 +46,6 @@ suite('Extension Test Suite', () => {
   })
 
   describe('dvc.setupWorkspace', () => {
-    it('should set dvc.pythonPath to the picked value when the user selects to pick a Python interpreter', async () => {
-      stub(DvcReader.prototype, 'version').rejects('do not initialize')
-      stub(Python, 'isPythonExtensionInstalled').returns(false)
-
-      const mockShowQuickPick = stub(window, 'showQuickPick')
-      const mockUri = Uri.file(
-        resolve('file', 'picked', 'path', 'to', 'python')
-      )
-      const mockPath = mockUri.fsPath
-      stub(window, 'showOpenDialog').resolves([mockUri])
-      const pythonChanged = configurationChangeEvent(
-        ConfigKey.PYTHON_PATH,
-        disposable
-      )
-
-      const venvQuickPickActive = quickPickInitialized(mockShowQuickPick, 0)
-
-      const quickPick = window.createQuickPick<QuickPickItemWithValue<string>>()
-      const mockCreateQuickPick = stub(window, 'createQuickPick').returns(
-        quickPick
-      )
-      const pickOneOrInputActive = new Promise(resolve => {
-        disposable.track(quickPick.onDidChangeActive(() => resolve(undefined)))
-      })
-
-      const setupWorkspaceWizard = commands.executeCommand(
-        RegisteredCommands.EXTENSION_SETUP_WORKSPACE
-      )
-
-      await venvQuickPickActive
-
-      const selectVenvAndInterpreter = selectQuickPickItem(1)
-      await selectVenvAndInterpreter
-
-      await pickOneOrInputActive
-      mockCreateQuickPick.restore()
-
-      const selectToFindInterpreter = selectQuickPickItem(1)
-      await selectToFindInterpreter
-
-      await pythonChanged
-
-      await setupWorkspaceWizard
-
-      expect(workspace.getConfiguration().get(ConfigKey.PYTHON_PATH)).to.equal(
-        mockPath
-      )
-    }).timeout(WEBVIEW_TEST_TIMEOUT)
-
     it('should initialize the extension when the cli is usable', async () => {
       stub(Python, 'isPythonExtensionInstalled').returns(true)
       const selectVirtualEnvWithPython = async (path: string) => {
@@ -296,24 +238,6 @@ suite('Extension Test Suite', () => {
 
       expect(mockCreateFileSystemWatcher).not.to.be.calledWithMatch('{}')
     }).timeout(25000)
-
-    it('should send an error telemetry event when setupWorkspace fails', async () => {
-      mockDuration(0)
-
-      const mockErrorMessage = 'NOPE'
-      stub(Setup, 'runWorkspace').rejects(new Error(mockErrorMessage))
-      const mockSendTelemetryEvent = stub(Telemetry, 'sendTelemetryEvent')
-
-      await expect(
-        commands.executeCommand(RegisteredCommands.EXTENSION_SETUP_WORKSPACE)
-      ).to.be.eventually.rejectedWith(Error)
-
-      expect(mockSendTelemetryEvent).to.be.calledWithExactly(
-        `errors.${RegisteredCommands.EXTENSION_SETUP_WORKSPACE}`,
-        { error: mockErrorMessage },
-        { duration: 0 }
-      )
-    }).timeout(WEBVIEW_TEST_TIMEOUT)
   })
 
   describe('dvc.stopRunningExperiment', () => {
@@ -351,75 +275,6 @@ suite('Extension Test Suite', () => {
       const showOutputSpy = spy(OutputChannel.prototype, 'show')
       await commands.executeCommand(RegisteredCommands.EXTENSION_SHOW_OUTPUT)
       expect(showOutputSpy).to.have.been.calledOnce
-    })
-  })
-
-  describe('dvc.checkCLICompatible', () => {
-    it('should call setup', async () => {
-      const mockSetup = stub(Setup, 'run').resolves(undefined)
-      await commands.executeCommand(
-        RegisteredCommands.EXTENSION_CHECK_CLI_COMPATIBLE
-      )
-      expect(mockSetup).to.have.been.calledOnce
-    })
-
-    it('should set the dvc.cli.incompatible context value', async () => {
-      stub(Watcher, 'createFileSystemWatcher').returns(undefined)
-      stub(DvcReader.prototype, 'expShow').resolves({
-        [EXPERIMENT_WORKSPACE_ID]: { baseline: {} }
-      })
-      stub(DvcReader.prototype, 'root').resolves('.')
-      stub(DvcReader.prototype, 'dataStatus').resolves({})
-      stub(DvcReader.prototype, 'plotsDiff').resolves({})
-      stub(GitReader.prototype, 'hasChanges').resolves(false)
-      stub(GitReader.prototype, 'listUntracked').resolves(new Set())
-      stub(Config.prototype, 'getPythonBinPath').resolves(join('python'))
-
-      stub(DvcReader.prototype, 'version')
-        .onFirstCall()
-        .resolves('1.0.0')
-        .onSecondCall()
-        .resolves(MIN_CLI_VERSION)
-        .onThirdCall()
-        .rejects(new Error('NO CLI HERE'))
-
-      const executeCommandSpy = spy(commands, 'executeCommand')
-      await commands.executeCommand(
-        RegisteredCommands.EXTENSION_CHECK_CLI_COMPATIBLE
-      )
-
-      expect(
-        executeCommandSpy,
-        'should set dvc.cli.incompatible to true if the version is incompatible'
-      ).to.be.calledWithExactly('setContext', 'dvc.cli.incompatible', true)
-      executeCommandSpy.resetHistory()
-
-      await commands.executeCommand(
-        RegisteredCommands.EXTENSION_CHECK_CLI_COMPATIBLE
-      )
-
-      expect(
-        executeCommandSpy,
-        'should set dvc.cli.incompatible to false if the version is compatible'
-      ).to.be.calledWithExactly('setContext', 'dvc.cli.incompatible', false)
-
-      const mockShowWarningMessage = stub(
-        window,
-        'showWarningMessage'
-      ).resolves(undefined)
-
-      await commands.executeCommand(
-        RegisteredCommands.EXTENSION_CHECK_CLI_COMPATIBLE
-      )
-
-      expect(
-        executeCommandSpy,
-        'should unset dvc.cli.incompatible if the CLI throws an error'
-      ).to.be.calledWithExactly('setContext', 'dvc.cli.incompatible', undefined)
-      expect(
-        mockShowWarningMessage,
-        'should warn the user if the CLI throws an error'
-      ).to.be.calledOnce
     })
   })
